refactor(blorktools): animation playback util
