+++ packages/blorktools/src/asset_debugger/util/css3d/css3d-prerender.js
@@ -1,5 +1,6 @@
 import { showStatus } from "../../modals/html-editor-modal/html-editor-modal";
 import { createMeshInfoPanel } from "../../modals/html-editor-modal/mesh-info-panel-util";
+import { updateMeshTexture } from "../animation-playback-util";
 import { 
     animationCaptureStartTime,
     animationDuration,
