+++ packages/blorktools/src/asset_debugger/util/custom-animation/animation-util.js
@@ -3,7 +3,7 @@ import { showStatus } from "../../modals/html-editor-modal/html-editor-modal";
 import { createMeshInfoPanel } from '../../modals/html-editor-modal/mesh-info-panel-util';
 import { createTextureFromIframe, injectUnifiedAnimationDetectionScript } from './html2canvas-util';
 
-const ANALYSIS_DURATION_MS = 30000; // 30 seconds - matches preRenderMaxDuration
+export const ANALYSIS_DURATION_MS = 30000; // 30 seconds - matches preRenderMaxDuration
 
 export let finalProgressAnimation = false;
 export let finalProgressStartTime = 0;
@@ -37,124 +37,23 @@ export let lastTextureUpdateTime = 0; // Exported for use in preview-util.js and
 export let animationPlaybackStartTime = 0; // When playback actually started
 export let animationCaptureStartTime = 0;  // When capture started (for offset calculations)
 // Active playback values
-let playbackStartTime = 0;
-let isPlaybackActive = false;
-
-/**
- * Start playback timing - called when preview should begin playing
- */
-export function startPlayback() {
-    playbackStartTime = Date.now();
-    isPlaybackActive = true;
-    console.log('Playback started at:', playbackStartTime);
-}
-
-/**
- * Stop playback timing
- */
-export function stopPlayback() {
-    isPlaybackActive = false;
-    console.log('Playback stopped');
-}
-
-/**
- * Get current frame based on elapsed playback time
- */
-export function getCurrentFrameForPlayback(playbackSpeed = 1.0, animationType = 'play') {
-    if (!isPlaybackActive || preRenderedFrames.length === 0) {
-        return preRenderedFrames.length > 0 ? preRenderedFrames[preRenderedFrames.length - 1] : null;
-    }
-    
-    const now = Date.now();
-    const playbackElapsed = now - playbackStartTime;
-    const adjustedElapsed = playbackElapsed * playbackSpeed;
-    
-    // Use the analysis duration as our animation duration
-    // This represents the full time period we analyzed for animation
-    const naturalDuration = isAnimationFinite && animationDuration > 0 ? 
-        animationDuration : ANALYSIS_DURATION_MS;
-    
-    let normalizedTime;
-    
-    switch (animationType) {
-        case 'play':
-            if (adjustedElapsed >= naturalDuration) {
-                setIsPreviewAnimationPaused(true);
-                return preRenderedFrames[preRenderedFrames.length - 1];
-            }
-            normalizedTime = adjustedElapsed / naturalDuration;
-            break;
-            
-        case 'loop':
-            normalizedTime = (adjustedElapsed % naturalDuration) / naturalDuration;
-            break;
-            
-        case 'bounce':
-            const cycle = Math.floor(adjustedElapsed / naturalDuration);
-            const position = (adjustedElapsed % naturalDuration) / naturalDuration;
-            normalizedTime = (cycle % 2 === 0) ? position : (1 - position);
-            break;
-            
-        default:
-            normalizedTime = (adjustedElapsed % naturalDuration) / naturalDuration;
-            break;
-    }
-    
-    const frameIndex = Math.min(
-        Math.floor(normalizedTime * preRenderedFrames.length),
-        preRenderedFrames.length - 1
-    );
-    
-    return preRenderedFrames[frameIndex];
-}
+export let playbackStartTime = 0;
+export let isPlaybackActive = false;
 
 /**
  * Reset all timing
  */
-export function resetPlaybackTiming() {
+export function resetPlaybackTimingState() {
     playbackStartTime = 0;
     isPlaybackActive = false;
 }
 
-
-
-export function resetPreRender() {
+export function resetPreRenderState() {
             preRenderedFrames = [];
             isAnimationFinite = false;
             preRenderingInProgress = false;
             finalProgressAnimation = false;
             finalProgressStartTime = 0;
-}
-
-/**
- * Update the mesh texture with the given texture
- * @param {THREE.Texture} texture - The texture to apply to the mesh
- * @param {THREE.Mesh} previewPlane - The mesh to update with the texture
- */
-export function updateMeshTexture(texture, previewPlane) {
-    if (!texture || !previewPlane || !previewPlane.material) return;
-    
-    let needsUpdate = false;
-    
-    if (Array.isArray(previewPlane.material)) {
-        previewPlane.material.forEach(material => {
-            if (material.map !== texture) {
-                material.map = texture;
-                needsUpdate = true;
-            }
-        });
-        
-        if (needsUpdate) {
-            previewPlane.material.forEach(material => {
-                material.needsUpdate = true;
-            });
-        }
-    } else {
-        if (previewPlane.material.map !== texture) {
-            previewPlane.material.map = texture;
-            previewPlane.material.needsUpdate = true;
-        }
-    }
 }
 
 /**
@@ -224,4 +123,12 @@ export function setAnimationPlaybackStartTime(incomingValue) {
 
 export function setAnimationCaptureStartTime(incomingValue) {
     animationCaptureStartTime = incomingValue;
+}
+
+export function setPlaybackStartTime(incomingValue) {
+    playbackStartTime = incomingValue;
+}
+
+export function setIsPlaybackActive(incomingValue) {
+    isPlaybackActive = incomingValue;
 }
